Minimize public surface for recover parity
