+++ rust/mcap/src/lib.rs
@@ -68,6 +68,7 @@
 
 pub mod read;
 pub mod records;
+pub mod recover;
 #[cfg(feature = "tokio")]
 pub mod tokio;
 pub mod write;
@@ -114,8 +115,6 @@ pub enum McapError {
     ConflictingChannels(String),
     #[error("Schema `{0}` has multiple records that don't match.")]
     ConflictingSchemas(String),
-    #[error("Chunk has multiple message index records for channel {0}")]
-    DuplicateMessageIndex(u16),
     #[error("Record parse failed")]
     Parse(#[from] binrw::Error),
     #[error("I/O error from writing, or reading a compression stream")]
